requirements.txt
customtkinter>=5.2.0
Pillow>=10.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pywin32>=306
schedule>=1.2.0
python-dotenv>=1.0.0
lxml>=5.0.0
comtypes>=1.1.0
pystray>=0.19.0
win10toast>=0.9


